+++ apps/cowswap-frontend/src/modules/affiliate/containers/AffiliatePartnerOnboard.tsx
@@ -3,6 +3,7 @@ import { ReactNode, useCallback } from 'react'
 import EARN_AS_AFFILIATE_ILLUSTRATION from '@cowprotocol/assets/images/earn-as-affiliate.svg'
 import { ButtonPrimary, ButtonSize } from '@cowprotocol/ui'
 import { useWalletDetails, useWalletInfo } from '@cowprotocol/wallet'
+import { useWalletChainId } from '@cowprotocol/wallet-provider'
 
 import { Trans } from '@lingui/react/macro'
 import styled from 'styled-components/macro'
@@ -33,13 +34,15 @@ import {
 
 export function AffiliatePartnerOnboard(): ReactNode {
   const { data: walletClient } = useWalletClient()
-  const { account, chainId } = useWalletInfo()
+  const { account } = useWalletInfo()
+  const chainId = useWalletChainId()
   const { walletName } = useWalletDetails()
   const onSelectNetwork = useOnSelectNetwork()
   const toggleWalletModal = useToggleWalletModal()
 
   const shouldHideNetworkSelector = useShouldHideNetworkSelector()
-  const isUnsupportedNetwork = !isSupportedPayoutsNetwork(chainId)
+  const onPayoutsChain = isSupportedPayoutsNetwork(chainId)
+  const shouldSwitchToPayoutsChain = !!account && !onPayoutsChain
   const isSignerAvailable = Boolean(walletClient)
   const partnerRewardAmount = getPartnerRewardAmountLabel()
   const triggerVolumeLabel = formatUsdCompact(getDefaultTriggerVolume())
@@ -68,12 +71,12 @@ export function AffiliatePartnerOnboard(): ReactNode {
               <Trans>Connect wallet</Trans>
             </ButtonPrimary>
           )}
-          {!!account && isUnsupportedNetwork && !shouldHideNetworkSelector && (
+          {shouldSwitchToPayoutsChain && !shouldHideNetworkSelector && (
             <ButtonPrimary buttonSize={ButtonSize.BIG} width={'320px'} onClick={onSwitchToMainnet}>
               <Trans>Switch to Ethereum</Trans>
             </ButtonPrimary>
           )}
-          {!!account && isUnsupportedNetwork && shouldHideNetworkSelector && (
+          {shouldSwitchToPayoutsChain && shouldHideNetworkSelector && (
             <WalletSwitchHint>
               <ButtonPrimary buttonSize={ButtonSize.BIG} disabled>
                 <Trans>Switch to Ethereum</Trans>
@@ -87,7 +90,7 @@ export function AffiliatePartnerOnboard(): ReactNode {
               </InlineNote>
             </WalletSwitchHint>
           )}
-          {!!account && !isUnsupportedNetwork && !isSignerAvailable && (
+          {!!account && onPayoutsChain && !isSignerAvailable && (
             <ButtonPrimary onClick={toggleWalletModal} data-testid="affiliate-unlock">
               <Trans>Become an affiliate</Trans>
             </ButtonPrimary>