+++ apps/cowswap-frontend/src/modules/injectedWidget/services/callWidgetHook.test.ts
@@ -1,12 +1,12 @@
 import { isInjectedWidget } from '@cowprotocol/common-utils'
 import { jotaiStore } from '@cowprotocol/core'
+import { getParentOrigin } from '@cowprotocol/iframe-transport'
 import { WidgetHookEvents, widgetIframeTransport, WidgetMethodsEmit } from '@cowprotocol/widget-lib'
 import type { WidgetMethodsEmitPayloadMap } from '@cowprotocol/widget-lib'
 
 import { callWidgetHook } from './callWidgetHook'
 
 import { injectedWidgetHooksEnabledAtom } from '../state/injectedWidgetHooksEnabledAtom'
-import { getParentOrigin } from '../utils/getParentOrigin.utils'
 
 jest.mock('@cowprotocol/common-utils', () => ({
   isInjectedWidget: jest.fn(),
@@ -18,7 +18,7 @@ jest.mock('@cowprotocol/core', () => ({
   },
 }))
 
-jest.mock('../utils/getParentOrigin.utils', () => ({
+jest.mock('@cowprotocol/iframe-transport', () => ({
   getParentOrigin: jest.fn(),
 }))
 
@@ -74,6 +74,7 @@ describe('callWidgetHook', () => {
 
     expect(mockListenToMessageFromWindow).toHaveBeenCalledWith(
       window,
+      expect.any(Window),
       'HOOK_RESULT',
       expect.any(Function),
       'http://localhost',
@@ -89,7 +90,7 @@ describe('callWidgetHook', () => {
       'http://localhost',
     )
 
-    const hookResultListener = mockListenToMessageFromWindow.mock.calls[0][2] as (payload: {
+    const hookResultListener = mockListenToMessageFromWindow.mock.calls[0][3] as (payload: {
       id: string
       result: boolean
     }) => void