+++ libs/common-const/src/cdn.ts
@@ -1,1 +1,1 @@
-export const COW_CDN = process.env['REACT_APP_COW_CDN'] || 'https://files.cow.finance'
+export const COW_CDN = process.env['REACT_APP_COW_CDN'] || 'https://files.cow.fi'