+++ libs/iframe-transport/src/iframeRpcProvider/WidgetEthereumProvider.ts
@@ -51,7 +51,7 @@ type RpcCallback = (error: any, response: any) => void
 const DEFAULT_TIMEOUT_MILLISECONDS = 600000
 
 const JSON_RPC_VERSION = '2.0'
-const DEFAULT_TARGET_ORIGIN = 'https://swap.cow.finance'
+const DEFAULT_TARGET_ORIGIN = 'https://swap.cow.fi'
 
 /**
  * Export the type information about the different events that are emitted.
@@ -190,6 +190,7 @@ export class WidgetEthereumProvider extends EventEmitter<IFrameEthereumProviderE
 
     iframeRpcProviderTransport.listenToMessageFromWindow(
       this.eventSource,
+      this.eventTarget,
       IframeRpcProviderEvents.PROVIDER_RPC_RESPONSE,
       (message) => {
         this.handleRpcRequests(message)
@@ -199,6 +200,7 @@ export class WidgetEthereumProvider extends EventEmitter<IFrameEthereumProviderE
 
     iframeRpcProviderTransport.listenToMessageFromWindow(
       this.eventSource,
+      this.eventTarget,
       IframeRpcProviderEvents.PROVIDER_ON_EVENT,
       (message) => {
         this.handleOnEvent(message)
@@ -208,6 +210,7 @@ export class WidgetEthereumProvider extends EventEmitter<IFrameEthereumProviderE
 
     iframeRpcProviderTransport.listenToMessageFromWindow(
       this.eventSource,
+      this.eventTarget,
       IframeRpcProviderEvents.SEND_PROVIDER_META_INFO,
       (message) => {
         this.providerMetaInfo = message