+++ libs/analytics/CHANGELOG.md
@@ -1,5 +1,50 @@
 # Changelog
 
+## [3.2.0](https://github.com/cowprotocol/cowswap/compare/analytics-v3.1.2...analytics-v3.2.0) (2026-04-16)
+
+
+### ✨ Features
+
+* bring cow.fi back (sdk) ([#7359](https://github.com/cowprotocol/cowswap/issues/7359)) ([7c19aee](https://github.com/cowprotocol/cowswap/commit/7c19aee0e0ec06ce9e2b657d79323975a62f6bd7))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 3.2.0
+    * @cowprotocol/common-hooks bumped to 3.2.0
+    * @cowprotocol/common-utils bumped to 3.3.0
+    * @cowprotocol/events bumped to 4.3.0
+
+## [3.1.2](https://github.com/cowprotocol/cowswap/compare/analytics-v3.1.1...analytics-v3.1.2) (2026-04-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 3.1.2
+    * @cowprotocol/common-hooks bumped to 3.1.2
+    * @cowprotocol/common-utils bumped to 3.2.1
+
+## [3.1.1](https://github.com/cowprotocol/cowswap/compare/analytics-v3.1.0...analytics-v3.1.1) (2026-04-15)
+
+
+### 🔧 Miscellaneous
+
+* merge main and resolve conflicts ([ed8f113](https://github.com/cowprotocol/cowswap/commit/ed8f113fc11c20e7e0fb84f9282abefa0693f8ab))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 3.1.1
+    * @cowprotocol/common-hooks bumped to 3.1.1
+    * @cowprotocol/common-utils bumped to 3.2.0
+    * @cowprotocol/events bumped to 4.2.1
+
 ## [3.1.0](https://github.com/cowprotocol/cowswap/compare/analytics-v3.0.5...analytics-v3.1.0) (2026-04-14)